Derive YadisServiceManager and OpenIDServiceEndpoint to make them (JSON) serializable
Both objects end up as data in a serialized object, when used with
django-openid-auth. Django switched to JSONSerializer that can no longer
serialize these objects (in contrast to the Picke-based serializer) and
hence throws an error. The easiest solution is to base both on dict
objects that can be easily be serialized by JSON, but it also requires
us to create corresponding getter/setter methods, plus ensure if we end
up with a plain dict, we "convert" that to the real object: The new
_from_dict functions.

Improve dictionary subclassing

Use a little metaprogramming to make the code less repetitive, while still maintaining the goal of quacking like a dictionary. In the case of the `YadisManager`, this requires changing the API a little bit to avoid interfering with `dict` built-in methods (particularly `__len__` and `__iter__`), but these seem little-used so it seems safe from the tests.

Also, fix a couple of bugs in parsing things `_from_dict`